+++ edgy/core/db/querysets/mixins/embedding.py
@@ -10,17 +10,46 @@ import sqlalchemy
 from edgy.core.db.context_vars import MODEL_GETATTR_BEHAVIOR
 from edgy.core.db.querysets.prefetch import Prefetch
 from edgy.core.db.querysets.types import EdgyEmbedTarget, EdgyModel, tables_and_models_type
+from edgy.core.db.relationships.utils import crawl_relationship
 from edgy.core.utils.concurrency import run_concurrently
 from edgy.core.utils.db import get_table_key_or_name
 from edgy.exceptions import QuerySetError
 
-from ..clauses import clean_path_to_crawl_result
+from .. import clauses as clauses_mod
 
 if TYPE_CHECKING:  # pragma: no cover
     from edgy import Model
+    from edgy.core.db.fields.base import BaseForeignKey
     from edgy.core.db.models.types import BaseModelType
     from edgy.core.db.querysets.prefetch import Prefetch
     from edgy.core.db.querysets.queryset import QuerySet
+
+
+def _apply_prefetches_helper(
+    *,
+    instance: BaseModelType,
+    prefetches: Sequence[Prefetch],
+    mapping: Mapping,
+    # if not provided we use row_prefix = "", required for embedding
+    tables_and_models: tables_and_models_type | None = None,
+) -> None:
+    """Apply prefetches to a specific model instance."""
+    for related in prefetches:
+        # Check for conflicting names early to prevent unexpected overwrites.
+        related.check_for_collision(model=instance)
+        # tables can be alias
+        row_prefix = (
+            f"{get_table_key_or_name(tables_and_models[related._forward_path_to_anchor][0])}_"
+            if related._forward_path_to_anchor and tables_and_models
+            else ""
+        )
+        model_key = related._anchor_model.create_model_key_from_raw_mapping(
+            mapping=mapping, prefix=row_prefix
+        )
+        # Ensure it is in the baked results.
+        related._baked_results.setdefault(model_key, [])
+        new_attr_name = related.to_attr.rsplit("__", 1)[-1]
+        object.__setattr__(instance, new_attr_name, list(related._baked_results[model_key]))
 
 
 class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
@@ -29,45 +58,13 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
     on a QuerySet like `prefetch_related` and `_embed_parent_in_result`.
     """
 
-    def _apply_prefetches_list(
-        self,
-        *,
-        instance: BaseModelType,
-        prefetches: Sequence[Prefetch],
-        mapping: Mapping,
-        prefix: str = "",
-        # if not provided we use row_prefix = "", required for embedding
-        tables_and_models: tables_and_models_type | None = None,
-    ) -> None:
-        """Apply prefetches to a specific model instance."""
-        for related in prefetches:
-            # Check for conflicting names early to prevent unexpected overwrites.
-            related.check_for_collision(model=instance)
-            reduced_prefix = prefix.removesuffix(related._forward_path_to_anchor).removesuffix(
-                "__"
-            )
-            # tables can be alias
-            row_prefix = (
-                f"{get_table_key_or_name(tables_and_models[reduced_prefix][0])}_"
-                if reduced_prefix and tables_and_models
-                else ""
-            )
-            model_key = related._target_model.create_model_key_from_raw_mapping(
-                mapping=mapping, prefix=row_prefix
-            )
-            # Ensure it is in the baked results.
-            related._baked_results.setdefault(model_key, [])
-            new_attr_name = related.to_attr.rsplit("__", 1)[-1]
-            object.__setattr__(instance, new_attr_name, list(related._baked_results[model_key]))
-
-    async def _apply_prefetches_self_and_select_related(
+    async def _apply_prefetches_self_and_related(
         self,
         *,
         instance: EdgyModel,
-        prefetches_dict: dict[str, list[Prefetch]],
+        prepared_prefetches: Sequence[Prefetch],
         mapping: Mapping,
         tables_and_models: tables_and_models_type | None,
-        seen: set[str],
     ) -> None:
         """Apply prefetches on select related branches.
 
@@ -77,55 +74,60 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
 
         Args:
             instance: The current EdgyModel instance.
-            prefetches_dict: A dictionary mapping relationship or embedding paths to lists of Prefetch objects.
+            prepared_prefetches: A Sequence of prepared Prefetch objects.
             mapping: The mapping object used for database operations.
             tables_and_models: Optional tables and models information.
-            seen: A set tracking already processed prefixes to prevent redundant fetches.
 
         Returns:
             None
         """
         self_queryset = cast("QuerySet[EdgyModel, EdgyEmbedTarget]", self)
+        seen: set[str] = set()
+        target_dict: dict[str, list[Prefetch]] = {}
+        resolve_anchor_dict: dict[str, list[Prefetch]] = {}
+        for prefetch in prepared_prefetches:
+            target_dict.setdefault(prefetch.forward_path, []).append(prefetch)
+            # FIXME: only add prefetches which are not complete yet
+            if prefetch.from_anchor:
+                resolve_anchor_dict.setdefault(prefetch.from_anchor, []).append(prefetch)
+
         token = MODEL_GETATTR_BEHAVIOR.set("passdown")
         try:
-            if "" not in seen and (prefetches_list := prefetches_dict.get("")):
-                seen.add("")
-                await run_concurrently(
-                    [prefetch._init_bake() for prefetch in prefetches_list],
-                    limit=1 if getattr(self_queryset.database, "force_rollback", False) else None,
-                )
-                self._apply_prefetches_list(
-                    instance=instance,
-                    mapping=mapping,
-                    tables_and_models=tables_and_models,
-                    prefetches=prefetches_list,
-                )
-            # we need only to check the automatically generated embedding.
-            for path in sorted(
-                self_queryset._select_related_embedding, key=lambda x: x.count("__"), reverse=True
-            ):
+            # we need only to check the prefetches dict
+            for path in sorted(target_dict, key=lambda x: len(x), reverse=True):
                 prefix = ""
                 current_instance: BaseModelType | None = instance
                 for part in path.split("__"):
                     prefix = f"{prefix}__{part}" if prefix else part
                     if prefix in seen:
                         continue
-                    current_instance = cast(
-                        "BaseModelType | None", getattr(current_instance, part, None)
-                    )
                     seen.add(prefix)
+                    old_instance = current_instance
+                    # empty part for empty pathes
+                    if part:
+                        current_instance = cast(
+                            "BaseModelType | None", getattr(current_instance, part, None)
+                        )
                     if current_instance is None:
                         break
-                    if prefetches_list := prefetches_dict.get(prefix):
+                    if (
+                        fk_field := cast(
+                            "BaseForeignKey | None", old_instance.meta.fields.get(part)
+                        )
+                    ) and fk_field.is_cross_db():
+                        raise NotImplementedError("db Traversal isn't implemented yet.")
+                        # await current_instance.load()
+                        # if part in resolve_anchor_dict:
+                        #    # TODO: use the anchor dictionary to resolve
+                    if prefetches_list := target_dict.get(prefix):
                         await run_concurrently(
                             [prefetch._init_bake() for prefetch in prefetches_list],
                             limit=1
                             if getattr(self_queryset.database, "force_rollback", False)
                             else None,
                         )
-                        self._apply_prefetches_list(
+                        _apply_prefetches_helper(
                             instance=current_instance,
-                            prefix=prefix,
                             mapping=mapping,
                             tables_and_models=tables_and_models,
                             prefetches=prefetches_list,
@@ -137,23 +139,26 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
     async def _embed_parent_in_result(
         self,
         result: None,
+        *,
+        prepared_prefetches: Sequence[Prefetch] | None = None,
         mapping: Mapping | None = None,
-        prefetches_dict: dict[str, list[Prefetch]] | None = None,
         tables_and_models: tables_and_models_type | None = None,
     ) -> tuple[None, None]: ...
     @overload
     async def _embed_parent_in_result(
         self,
         result: EdgyModel | Awaitable[EdgyModel],
+        *,
+        prepared_prefetches: Sequence[Prefetch] | None = None,
         mapping: Mapping | None = None,
-        prefetches_dict: dict[str, list[Prefetch]] | None = None,
         tables_and_models: tables_and_models_type | None = None,
     ) -> tuple[EdgyModel, EdgyEmbedTarget]: ...
     async def _embed_parent_in_result(
         self,
         result: EdgyModel | Awaitable[EdgyModel] | None,
+        *,
+        prepared_prefetches: Sequence[Prefetch] | None = None,
         mapping: Mapping | None = None,
-        prefetches_dict: dict[str, list[Prefetch]] | None = None,
         tables_and_models: tables_and_models_type | None = None,
     ) -> tuple[EdgyModel, EdgyEmbedTarget] | tuple[None, None]:
         """
@@ -164,14 +169,25 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
             result = await result
         if result is None:
             return None, None
-        seen_prefixes: set[str] = set()
-        if mapping is not None and prefetches_dict:
-            await self._apply_prefetches_self_and_select_related(
-                instance=result,
-                prefetches_dict=prefetches_dict,
+        if mapping is not None and prepared_prefetches:
+            if tables_and_models is None:
+                tables_and_models = (await self_queryset.as_select_with_tables())[1]
+            # apply embedding
+            prefetched_instance = result
+            if self_queryset.embed_parent_filters and self_queryset.embed_parent_filters[0]:
+                token = MODEL_GETATTR_BEHAVIOR.set("coro")
+                try:
+                    for part in self_queryset.embed_parent_filters[0].split("__"):
+                        prefetched_instance = getattr(prefetched_instance, part)
+                        if isawaitable(prefetched_instance):
+                            prefetched_instance = await prefetched_instance
+                finally:
+                    MODEL_GETATTR_BEHAVIOR.reset(token)
+            await self._apply_prefetches_self_and_related(
+                instance=prefetched_instance,
+                prepared_prefetches=prepared_prefetches,
                 mapping=mapping,
                 tables_and_models=tables_and_models,
-                seen=seen_prefixes,
             )
         if not self_queryset.embed_parent:
             return result, cast("EdgyEmbedTarget", result)
@@ -184,20 +200,6 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
                 new_result = getattr(new_result, part)
                 if isawaitable(new_result):
                     new_result = await new_result
-                # if (
-                #     tables_and_models is not None
-                #     and mapping is not None
-                #     and prefix not in seen_prefixes
-                #     and prefetches_dict
-                #     and (prefetches_list := prefetches_dict.get(prefix))
-                # ):
-                #     await self._apply_prefetches_list(
-                #         instance=new_result,
-                #         mapping=mapping,
-                #         prefetches=prefetches_list,
-                #         # not in select related, don't provide prefix
-                #     )
-                # seen_prefixes.add(prefix)
         finally:
             MODEL_GETATTR_BEHAVIOR.reset(token)
         if self_queryset.embed_parent[1]:
@@ -206,7 +208,7 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
 
     def _prepare_prefetches_for_rows(
         self, rows: Sequence[sqlalchemy.Row], tables_and_models: tables_and_models_type
-    ) -> dict[str, list[Prefetch]]:
+    ) -> Sequence[Prefetch]:
         """
         Builds the Prefetch objects for a given batch of results.
         This is the *prefetch building* half of the original _handle_batch.
@@ -223,57 +225,68 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
             QuerySetError: If a prefetch path is invalid (e.g., unidirectional).
         """
         self_queryset = cast("QuerySet[EdgyModel, EdgyEmbedTarget]", self)
-        prepared_prefetches: dict[str, list[Prefetch]] = {}
-        seen_prefetches: set[tuple[None | int, str, str, str]] = set()
+        prepared_prefetches: dict[tuple[str, str], Prefetch] = {}
+        seen_prefetches: set[tuple[str, str, str]] = set()
 
         for prefetch in self_queryset._prefetch_related:
             compare_tuple = (
-                id(prefetch.queryset) if prefetch.queryset is not None else None,
                 prefetch.related_name,
-                prefetch.from_anchor,
+                prefetch.forward_path,
                 prefetch.to_attr,
             )
             if compare_tuple in seen_prefetches:
                 continue
             else:
                 seen_prefetches.add(compare_tuple)
-            target_crawl_result = clean_path_to_crawl_result(
-                self_queryset.model_class,
-                prefetch.to_attr,
-                embed_parent=self.embed_parent_filters,
-                # allow_crossing_db=True,
-                # no_operator=True,
-            )
-            anchor_crawl_result = clean_path_to_crawl_result(
+            anchor_crawl_result = crawl_relationship(
                 self_queryset.model_class,
                 prefetch.from_anchor,
-                embed_parent=self.embed_parent_filters,
-                model_database=self.database,
-                path_to_field=False,
+                embed_parent=self_queryset.embed_parent_filters,
+                model_database=self_queryset.database,
                 # allow_crossing_db=True,
-                # traverse_last=True,
-                # no_operator=True,
+                traverse_last=True,
             )
+            if anchor_crawl_result.field_name or anchor_crawl_result.operator:
+                raise QuerySetError(
+                    detail=f"`from_anchor` points to non-relation field: `{anchor_crawl_result.field_name}`."
+                )
+
             if anchor_crawl_result.cross_db_remainder:
                 raise NotImplementedError(
                     "Cannot prefetch from other db yet. Maybe in future this feature will be added."
                 )
 
-            prefetch_crawl_result = clean_path_to_crawl_result(
+            prefetch_crawl_result = crawl_relationship(
                 anchor_crawl_result.model_class,
                 prefetch.related_name,
-                path_to_field=False,
-                embed_parent=self.embed_parent_filters,
-                # traverse_last=True,
-                # no_operator=True,
+                traverse_last=True,
             )
+            if prefetch_crawl_result.field_name or anchor_crawl_result.operator:
+                raise QuerySetError(
+                    detail=f"`related_name` path points to non-relation field: `{prefetch_crawl_result.field_name}`."
+                )
             if prefetch_crawl_result.cross_db_remainder:
                 raise NotImplementedError(
                     "Cannot prefetch from other db yet. Maybe in future this feature will be added."
                 )
             if prefetch_crawl_result.reverse_path is False:
                 raise QuerySetError(
-                    detail=("Creating a reverse path is not possible, unidirectional fields used.")
+                    detail=(
+                        "Creating a reverse path to `from_anchor` is not possible, unidirectional fields were used."
+                    )
+                )
+
+            target_crawl_result = crawl_relationship(
+                anchor_crawl_result.model_class,
+                prefetch.to_attr,
+                traverse_last=True,
+            )
+            if target_crawl_result.reverse_path is False:
+                raise QuerySetError(
+                    detail=(
+                        "Creating a reverse path from `to_attr` to `from_anchor` is not possible, "
+                        "unidirectional fields were used."
+                    )
                 )
 
             prefetch.check_for_collision(anchor_crawl_result.model_class)
@@ -302,20 +315,26 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
                 prefetch_queryset = prefetch_crawl_result.model_class.query.local_or(*clauses)
             else:
                 prefetch_queryset = prefetch_queryset.local_or(*clauses)
-
-            prefetch_queryset = prefetch_queryset.select_related(
-                prefetch_crawl_result.reverse_path
-            )
+            # just add the resolved path to _select_related_embedding, so get less cruft
+            prefetch_queryset._select_related_embedding.add(prefetch_crawl_result.reverse_path)
             # the assigned queryset has an empty cache
             new_prefetch.queryset = prefetch_queryset
+            # don't calculate twice
             new_prefetch.forward_path = prefetch.forward_path
-            new_prefetch._forward_path_to_anchor = prefetch_crawl_result.forward_path
+            new_prefetch._forward_path_to_anchor = anchor_crawl_result.forward_path
             new_prefetch._reverse_path_to_anchor = prefetch_crawl_result.reverse_path
             new_prefetch._baking_finished = asyncio.Event()
-            new_prefetch._target_model = cast("type[Model]", target_crawl_result.model_class)
+            new_prefetch._anchor_model = cast("type[Model]", anchor_crawl_result.model_class)
             new_prefetch._baked_results = {}
-            prepared_prefetches.setdefault(new_prefetch.forward_path, []).append(new_prefetch)
-        return prepared_prefetches
+            prefetch_key = (new_prefetch.forward_path, new_prefetch.to_attr.rsplit("__", 1)[-1])
+            if prefetch_key in prepared_prefetches:
+                raise QuerySetError(
+                    f"Conflicting prefetches to_attr='{prefetch_key[1]}' for path=`{prefetch_key[1]}` "
+                    f"'on {prefetch_crawl_result.model_class.__name__}"
+                )
+
+            prepared_prefetches[prefetch_key] = new_prefetch
+        return list(prepared_prefetches.values())
 
     def prefetch_related(self, *prefetch: Prefetch) -> QuerySet[EdgyModel, EdgyEmbedTarget]:
         """
@@ -359,9 +378,16 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
         if queryset.embed_parent and queryset.embed_parent[0]:
             # parsed later
             select_pathes.add(queryset.embed_parent[0])
-        # now add the forward pathes
+        # now add the forward pathes, we need to resolve them because embed_parent is resolved
         select_pathes.update(
-            prefetch.forward_path
+            clauses_mod.clean_path_to_crawl_result(
+                self_queryset.model_class,
+                prefetch.forward_path,
+                model_database=self_queryset.database,
+                embed_parent=self_queryset.embed_parent_filters,
+                # we need this as we have no field name here
+                path_to_field=False,
+            ).forward_path
             for prefetch in queryset._prefetch_related
             if prefetch.forward_path
         )
@@ -409,7 +435,14 @@ class EmbeddingMixin(Generic[EdgyModel, EdgyEmbedTarget]):
 
             queryset._update_select_related_weak(
                 (
-                    prefetch.forward_path
+                    clauses_mod.clean_path_to_crawl_result(
+                        self_queryset.model_class,
+                        prefetch.forward_path,
+                        model_database=self_queryset.database,
+                        embed_parent=self_queryset.embed_parent_filters,
+                        # we need this as we have no field name here
+                        path_to_field=False,
+                    ).forward_path
                     for prefetch in queryset._prefetch_related
                     if prefetch.forward_path
                 ),